+++ src/data/damage-roll/critical-specialization/sword.ts
@@ -1,13 +1,16 @@
 import { AssistantAction } from "action.ts";
 import { EffectSource } from "foundry-pf2e";
 import { AssistantMessage } from "message.ts";
+import { Utils } from "utils.ts";
 
 export const actions: AssistantAction[] = [
     {
         trigger: "damage-roll",
         predicate: ["check:outcome:critical-success", "critical-specialization", "item:group:sword"],
         process: async (message: AssistantMessage) => {
-            if (!message.speaker?.actor || !message.target?.actor) return;
+            if (!message.speaker?.actor) return;
+            if (!message.target?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             game.assistant.socket.addEmbeddedItem(
                 message.target.actor,
@@ -27,8 +30,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },