+++ src/data/skill-check/action/demoralize.ts
@@ -1,6 +1,7 @@
 import { AssistantAction } from "action.ts";
 import { EffectSource } from "foundry-pf2e";
 import { AssistantMessage } from "message.ts";
+import { Utils } from "utils.ts";
 
 export const actions: AssistantAction[] = [
     {
@@ -9,6 +10,7 @@ export const actions: AssistantAction[] = [
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
             if (!message.target?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             const immunity = message.target.actor.itemTypes.effect.find(
                 (effect) =>
@@ -45,8 +47,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },
@@ -60,6 +62,7 @@ export const actions: AssistantAction[] = [
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
             if (!message.target?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             const immunity = message.target.actor.itemTypes.effect.find(
                 (effect) =>
@@ -96,8 +99,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },
@@ -111,6 +114,7 @@ export const actions: AssistantAction[] = [
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
             if (!message.target?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             game.assistant.socket.addEmbeddedItem(
                 message.target.actor,
@@ -130,8 +134,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },