+++ src/data/skill-check/action/feint.ts
@@ -1,6 +1,7 @@
 import { AssistantAction } from "action.ts";
 import { EffectPF2e, EffectSource } from "foundry-pf2e";
 import { AssistantMessage } from "message.ts";
+import { Utils } from "utils.ts";
 
 export const actions: AssistantAction[] = [
     {
@@ -9,6 +10,7 @@ export const actions: AssistantAction[] = [
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
             if (!message.target?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             game.assistant.socket.addEmbeddedItem(
                 message.speaker.actor,
@@ -28,8 +30,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },
@@ -43,6 +45,7 @@ export const actions: AssistantAction[] = [
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
             if (!message.target?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             game.assistant.socket.addEmbeddedItem(
                 message.speaker.actor,
@@ -62,8 +65,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },
@@ -77,6 +80,7 @@ export const actions: AssistantAction[] = [
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
             if (!message.target?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             let effect = await fromUuid<EffectPF2e>(
                 "Compendium.pf2e-assistant.pf2e-assistant-effects.Item.9f4YzsgcAs3A5Xra",
@@ -97,8 +101,8 @@ export const actions: AssistantAction[] = [
                                 token: message.speaker.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },