+++ src/data/skill-check/class-feature/swashbuckler/panache.ts
@@ -1,6 +1,7 @@
 import { AssistantAction } from "action.ts";
 import { EffectSource } from "foundry-pf2e";
 import { AssistantMessage } from "message.ts";
+import { Utils } from "utils.ts";
 
 export const actions: AssistantAction[] = [
     {
@@ -12,6 +13,7 @@ export const actions: AssistantAction[] = [
         ],
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             game.assistant.socket.addEmbeddedItem(
                 message.speaker.actor,
@@ -31,8 +33,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target?.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },
@@ -45,6 +47,7 @@ export const actions: AssistantAction[] = [
         predicate: ["item:trait:bravado", { not: "self:effect:panache" }, "check:outcome:failure"],
         process: async (message: AssistantMessage) => {
             if (!message.speaker?.actor) return;
+            if (!Utils.isInstanceOf(message.roll, "CheckRoll")) return;
 
             game.assistant.socket.addEmbeddedItem(
                 message.speaker.actor,
@@ -64,8 +67,8 @@ export const actions: AssistantAction[] = [
                                 token: message.target?.token?.uuid ?? null,
                             },
                             roll: {
-                                degreeOfSuccess: message.checkRoll?.degreeOfSuccess,
-                                total: message.checkRoll?.total ?? null,
+                                degreeOfSuccess: message.roll?.degreeOfSuccess,
+                                total: message.roll?.total ?? null,
                             },
                         },
                     },